+++ src/Ydb.Sdk/src/Ado/Session/PoolingSession.cs
@@ -1,7 +1,9 @@
+using System.Diagnostics;
 using Microsoft.Extensions.Logging;
 using Ydb.Query;
 using Ydb.Query.V1;
 using Ydb.Sdk.Ado.Internal;
+using Ydb.Sdk.Tracing;
 using CommitTransactionRequest = Ydb.Query.CommitTransactionRequest;
 using TransactionControl = Ydb.Query.TransactionControl;
 
@@ -62,12 +64,16 @@ internal class PoolingSession : PoolingSessionBase<PoolingSession>
         return Driver.ServerStreamCall(QueryService.ExecuteQueryMethod, request, settings);
     }
 
-    public override async Task CommitTransaction(string txId, CancellationToken cancellationToken = default)
+    public override async Task CommitTransaction(
+        string txId,
+        Activity? dbActivity,
+        CancellationToken cancellationToken
+    )
     {
         var response = await Driver.UnaryCall(
             QueryService.CommitTransactionMethod,
             new CommitTransactionRequest { SessionId = SessionId, TxId = txId },
-            new GrpcRequestSettings { CancellationToken = cancellationToken, NodeId = NodeId }
+            new GrpcRequestSettings { CancellationToken = cancellationToken, NodeId = NodeId, DbActivity = dbActivity }
         );
 
         if (response.Status.IsNotSuccess())
@@ -76,12 +82,16 @@ internal class PoolingSession : PoolingSessionBase<PoolingSession>
         }
     }
 
-    public override async Task RollbackTransaction(string txId, CancellationToken cancellationToken = default)
+    public override async Task RollbackTransaction(
+        string txId,
+        Activity? dbActivity,
+        CancellationToken cancellationToken
+    )
     {
         var response = await Driver.UnaryCall(
             QueryService.RollbackTransactionMethod,
             new RollbackTransactionRequest { SessionId = SessionId, TxId = txId },
-            new GrpcRequestSettings { CancellationToken = cancellationToken, NodeId = NodeId }
+            new GrpcRequestSettings { CancellationToken = cancellationToken, NodeId = NodeId, DbActivity = dbActivity }
         );
 
         if (response.Status.IsNotSuccess())
@@ -111,7 +121,10 @@ internal class PoolingSession : PoolingSessionBase<PoolingSession>
 
     internal override async Task Open(CancellationToken cancellationToken)
     {
-        var requestSettings = new GrpcRequestSettings { CancellationToken = cancellationToken };
+        using var dbActivity = YdbActivitySource.StartActivity("ydb.CreateSession");
+
+        var requestSettings = new GrpcRequestSettings
+            { CancellationToken = cancellationToken, DbActivity = dbActivity };
 
         if (!_disableServerBalancer)
         {
@@ -207,7 +220,15 @@ internal class PoolingSession : PoolingSessionBase<PoolingSession>
             }
         }, cancellationToken);
 
-        await completeTask.Task;
+        try
+        {
+            await completeTask.Task;
+        }
+        catch (Exception e)
+        {
+            dbActivity?.SetException(e);
+            throw;
+        }
     }
 
     internal override async Task DeleteSession()