+++ src/Ydb.Sdk/src/Ado/Session/RetryableSession.cs
@@ -1,6 +1,8 @@
+using System.Diagnostics;
 using Ydb.Query;
 using Ydb.Sdk.Ado.Internal;
 using Ydb.Sdk.Ado.RetryPolicy;
+using Ydb.Sdk.Tracing;
 
 namespace Ydb.Sdk.Ado.Session;
 
@@ -34,11 +36,17 @@ internal class RetryableSession : ISession
             new InMemoryServerStream(_sessionSource, _retryPolicyExecutor, query, parameters, settings));
     }
 
-    public Task CommitTransaction(string txId, CancellationToken cancellationToken = default) =>
-        throw NotSupportedTransaction;
+    public Task CommitTransaction(
+        string txId,
+        Activity? dbActivity = null,
+        CancellationToken cancellationToken = default
+    ) => throw NotSupportedTransaction;
 
-    public Task RollbackTransaction(string txId, CancellationToken cancellationToken = default) =>
-        throw NotSupportedTransaction;
+    public Task RollbackTransaction(
+        string txId,
+        Activity? dbActivity = null,
+        CancellationToken cancellationToken = default
+    ) => throw NotSupportedTransaction;
 
     public void OnNotSuccessStatusCode(StatusCode code)
     {
@@ -82,6 +90,12 @@ internal sealed class InMemoryServerStream : IServerStream<ExecuteQueryResponseP
         {
             using var session = await _sessionSource.OpenSession(ct);
 
+            using var dbActivity = YdbActivitySource.StartActivity("ydb.ExecuteQuery");
+
+            _settings.DbActivity = dbActivity is { IsAllDataRequested: true }
+                ? dbActivity.SetTag("ydb.execute.in_memory", true)
+                : dbActivity;
+
             try
             {
                 var responses = new List<ExecuteQueryResponsePart>();
@@ -104,6 +118,7 @@ internal sealed class InMemoryServerStream : IServerStream<ExecuteQueryResponseP
             catch (YdbException e)
             {
                 session.OnNotSuccessStatusCode(e.Code);
+                dbActivity?.SetException(e);
                 throw;
             }
         }, cancellationToken);